refactor: extract the save pipeline and the admin list table
Completes the split. Documentate_Documents keeps the post type, the hooks and
the document-type lock; the four jobs it used to do itself now live in
collaborators, the way wp-decker arranges its event CPT.

  Documentate_Document_Meta_Boxes   57 methods   draws the editor
  Documentate_Document_Meta_Saver   27 methods   writes meta, composes content
  Documentate_Document_Admin_List   14 methods   filters, columns, views
  Documentate_Documents            560 lines     registration and hooks

Both boundaries were clean before the move: neither block reached anything
that stayed behind, and the two share nothing with each other.

The two classes are hooked differently, on purpose. The admin list registers
its own hooks, which is what wp-decker does and what takes eight public
methods off the CPT - enough to resolve TooManyPublicMethods, which no amount
of moving private methods could. The saver keeps its hooks on the CPT and
delegates, matching the metabox class, because 27 tests call
filter_post_data_compose_content() and save_meta_boxes() by name and the
public-method budget did not need them.

Two mistakes worth recording. The extraction rewrote $this-> calls but not
self:: ones, so three references in the saver silently retargeted to the new
class and 30 tests failed with "undefined method
Documentate_Document_Meta_Saver::parse_structured_content". Four more failed
downstream of those, which is why "Documentate_Conversion_Manager not found"
appeared and then vanished once the real cause was fixed - it was never a
loading problem. Separately, injecting the property docblocks put one between
an existing docblock and its declaration, orphaning it; PHPCS caught it.

Two tests asserted the hook callbacks by object identity, which is exactly
what moving the admin list changes. They now instantiate the class and assert
it registers its own hooks, which is the behaviour worth pinning.

Verified by rendering the editor, composing a saved post_content and building
the list columns and views, then diffing all three against the pre-split code:
identical.

Documentate_Documents now has no PHPMD alerts at all, down from four.
Plugin-wide 52 to 50: Meta_Boxes and Meta_Saver each carry an
ExcessiveClassComplexity of their own, 204 and 138 against a threshold of 100.
