+++ tests/unit/includes/custom-post-types/DocumentateDocumentsHelpersTest.php
@@ -12,6 +12,20 @@
 class DocumentateDocumentsHelpersTest extends WP_UnitTestCase {
 
 	/**
+	 * Field persistence.
+	 *
+	 * @var Documentate_Document_Meta_Saver
+	 */
+	private $meta_saver;
+
+	/**
+	 * Admin list table behaviour.
+	 *
+	 * @var Documentate_Document_Admin_List
+	 */
+	private $admin_list;
+
+	/**
 	 * Metabox renderer, which now owns the rendering internals.
 	 *
 	 * @var Documentate_Document_Meta_Boxes
@@ -33,6 +47,27 @@ class DocumentateDocumentsHelpersTest extends WP_UnitTestCase {
 
 		$this->documents = new Documentate_Documents();
 		$this->meta_boxes = new Documentate_Document_Meta_Boxes();
+		$this->meta_saver = new Documentate_Document_Meta_Saver();
+		$this->admin_list = new Documentate_Document_Admin_List();
+	}
+
+	/**
+	 * Find whichever collaborator declares a method.
+	 *
+	 * The CPT was split into a renderer, a saver and an admin-list class, so a
+	 * helper under test now lives on one of four objects.
+	 *
+	 * @param string $name Method name.
+	 * @return object
+	 */
+	private function owner_of( $name ) {
+		foreach ( array( $this->documents, $this->meta_boxes, $this->meta_saver, $this->admin_list ) as $candidate ) {
+			if ( method_exists( $candidate, $name ) ) {
+				return $candidate;
+			}
+		}
+
+		$this->fail( 'No collaborator declares ' . $name );
 	}
 
 	/**
@@ -43,7 +78,7 @@ class DocumentateDocumentsHelpersTest extends WP_UnitTestCase {
 	 * @return mixed
 	 */
 	private function invoke( $name, array $args = array() ) {
-		$target = method_exists( $this->documents, $name ) ? $this->documents : $this->meta_boxes;
+		$target = $this->owner_of( $name );
 		$method = ( new ReflectionClass( $target ) )->getMethod( $name );
 		$method->setAccessible( true );
